feat: sistema de múltiplos atlas, tradução pt-BR e correções de estabilidade
- Motor de Arte: implementado suporte a múltiplos álbuns (Texture2DArray) no ImageManager.cs e Shaders, permitindo carregar até 8192 imagens (superando o limite de 2048 do DirectX 11).
- Fallbacks de Imagem: adicionada busca robusta de artes de criaturas com suporte a prefixos (FISH_, DIVINE_, etc.), limpeza de sufixos e normalização de nomes.
- Shaders: atualizados CreatureSprite e ToplitSprite para seleção dinâmica de atlas via índice de camada.
- Localização: implementada tradução completa para Português (Brasil) e corrigida a interatividade do seletor de idiomas na UI.
- Estabilidade: corrigido erro de IndexOutOfRangeException no TextureScale.cs e removido uso de APIs obsoletas do Unity (isNetworkError, forceModuleActive).
- Bugfix: resolvida a falha de "itens amarelos" corrigindo o tipo de propriedade de shader (float vs int).

+++ Assets/Scripts/BuildingMaterialSetter.cs
@@ -16,7 +16,7 @@ public class BuildingMaterialSetter : MonoBehaviour
             var renderer = part.GetComponent<MeshRenderer>();
             MaterialPropertyBlock prop = new MaterialPropertyBlock();
             prop.SetColor("_MatColor", set.color);
-            prop.SetFloat("_MatIndex", set.patternIndex);
+            prop.SetInt("_MatIndex", set.patternIndex);
             renderer.SetPropertyBlock(prop);
         }
         foreach (var part in FindObjectsOfType<ItemModel>())
