+++ Assets/Scripts/MapGen/Items/ItemImprovement.cs
@@ -17,8 +17,8 @@ public class ItemImprovement : MonoBehaviour
     internal void UpdateImprovement(RemoteFortressReader.ItemImprovement improvement)
     {
         Color matColor = ContentLoader.GetColor(improvement.material);
-        float textureIndex = ContentLoader.GetPatternIndex(improvement.material);
-        float shapeIndex = ContentLoader.GetShapeIndex(improvement.material);
+        int textureIndex = ContentLoader.GetPatternIndex(improvement.material);
+        int shapeIndex = ContentLoader.GetShapeIndex(improvement.material);
         image = improvement.image;
 
         if (actualModel != null)
@@ -71,8 +71,8 @@ public class ItemImprovement : MonoBehaviour
 
         MaterialPropertyBlock prop = new MaterialPropertyBlock();
         prop.SetColor("_MatColor", matColor);
-        prop.SetFloat("_MatIndex", textureIndex);
-        prop.SetFloat("_ShapeIndex", shapeIndex);
+        prop.SetInt("_MatIndex", textureIndex);
+        prop.SetInt("_ShapeIndex", shapeIndex);
         meshRenderer.SetPropertyBlock(prop);
         gameObject.SetActive(true);
     }
